fix(edit): re-anchor before reporting stale capability bounds
The SDK replace_range applicator rejected an out-of-range capability before attempting relocation, so it refused tokens the agent-runtime transaction path would have recovered. reanchorCapabilityRange now runs first; only on failure does the applicator choose between the verbatim out-of-bounds diagnostic and the existing re-anchor failure cause. No production change was needed in common/src/util/line-coordinates.ts.

Split read-files-edit-state.test.ts (8,774 lines) by moving the capability-authorization and replace_range cases into read-files-capability-authorization.test.ts and extracting shared scaffolding into a non-test read-files-edit-state-fixtures.ts. Pure relocation: no assertion, name, or fixture value changed, and the 120-test total is preserved.

Fold in four advisory cleanups: a shared DIAGNOSTIC_WRITE_TARGET_TOKEN backs both diagnostic redirect patterns so they cannot drift apart; JobRegistry.clear() resolves pending wait() promises with undefined instead of leaking them, which previously hung any caller awaiting across a clear; FilesystemAuthority routes all three issue* methods through one recordCanonicalReceipt so the receipt log stays bounded; and the line-coordinates budget tests derive their fixtures from MAX_REANCHOR_SCAN_LINE_PRODUCT with a just-under-budget sibling.

common/knowledge.md is refreshed per the pre-push memory-drift guard and now documents the clear() waiter-resolution contract.
